Update number text field on compose to handle decimal points and nullable numbers.

+++ lib-ui-compose/src/commonMain/kotlin/com/ustadmobile/libuicompose/view/clazzassignment/UstadAssignmentSubmissionHeader.kt
@@ -19,7 +19,7 @@ import dev.icerock.moko.resources.compose.stringResource
 fun LazyListScope.UstadAssignmentSubmissionStatusHeaderItems(
     submissionStatus: Int?,
     averageMark: AverageCourseAssignmentMark?,
-    maxPoints: Int = 0,
+    maxPoints: Float = 0f,
     submissionPenaltyPercent: Int = 0,
 ) {
     item(key = "submission_status") {
